# Fix join pruning performance