@sumedhpathak @ozgune Could either of you review this soon-ish? Or assign somebody to do that? I think both fixes are fairly simple, the first being nearly trivial.

The first commit gets plan time for the query in issue #607 goes from 8764 ms to 2889 ms, the second from 2889 ms to 106 ms.

Fixes #607